+++ python/target_selection/cartons/mwm_halo_gaia_dr3.py
@@ -165,7 +165,7 @@ Link to paper: https://arxiv.org/abs/2302.02611 and
 Zenodo: https://doi.org/10.5281/zenodo.7599789
 Use catalogdb.xpfeh_gaia_dr3.
 
-Cut on BP<17, teff_xgboost < 5500, logg_xgboost < 4,
+Cut on BP<17, teff_xgboost < 5500, logg_xgboost < 4, mh_xgboost < -1.0
 w1mpro (from AllWise) absolute magnitude.
 
 Specifically:
@@ -175,6 +175,7 @@ with teff_xgboost on right hand side)
 BP < 17
 logg_xgboost < 4.0
 teff_xgboost < 5500
+mh_xgboost < -1.0
 M_W1 > -0.3 - 0.006 * (5500 - teff_xgboost)
 M_W1 > -0.01 * (5300 - teff_xgboost)
 
@@ -190,7 +191,7 @@ Metadata:
 Priority: three levels
 2100 if mh_xgboost <= -2.0 (TBA pending A/B test)
 2970 if -2.0 < mh_xgboost <= -1.5
-6090 if mh_xgboost > -1.5
+6090 if -1.5 < mh_xgboost <= -1.0
 Cadence: bright_1x1
 Instrument: BOSS for G>13, APOGEE for G<13
 can_offset = True
@@ -208,6 +209,12 @@ Lead contact: Alexander Ji, Rene Andrae
 
     def build_query(self, version_id, query_region=None):
 
+        # We use the lower case variable name m_w1 so that the
+        # corresponding column name is lower case.
+        m_w1 = AllWise.w1mpro + 5 * peewee.fn.log10(Gaia_DR3.parallax / 100)
+
+        # w1mpro is of PostgreSQL numeric(5,3) type
+        # Hence, below we cast it to real so that astropy can handle it.
         query = (CatalogToGaia_DR3
                  .select(CatalogToGaia_DR3.catalogid,
                          Gaia_DR3.source_id,
@@ -215,10 +222,12 @@ Lead contact: Alexander Ji, Rene Andrae
                          Gaia_DR3.dec.alias('gaia_dr3_dec'),
                          Gaia_DR3.phot_bp_mean_mag,
                          Gaia_DR3.phot_g_mean_mag,
+                         Gaia_DR3.parallax,
                          Xpfeh_gaia_dr3.logg_xgboost,
                          Xpfeh_gaia_dr3.teff_xgboost,
                          Xpfeh_gaia_dr3.mh_xgboost,
-                         AllWise.w1mpro)
+                         AllWise.w1mpro.cast('real'),
+                         m_w1.alias('m_w1'))
                  .join(Gaia_DR3, on=(CatalogToGaia_DR3.target_id == Gaia_DR3.source_id))
                  .join(Xpfeh_gaia_dr3,
                        on=(Gaia_DR3.source_id == Xpfeh_gaia_dr3.source_id))
@@ -229,9 +238,15 @@ Lead contact: Alexander Ji, Rene Andrae
                        on=(CatalogToAllWise.target_id == AllWise.cntr))
                  .where(CatalogToGaia_DR3.version_id == version_id,
                         CatalogToGaia_DR3.best >> True,
+                        CatalogToAllWise.version_id == version_id,
+                        CatalogToAllWise.best >> True,
                         Gaia_DR3.phot_bp_mean_mag < 17,
+                        Gaia_DR3.parallax > 0,
                         Xpfeh_gaia_dr3.logg_xgboost < 4.0,
-                        Xpfeh_gaia_dr3.teff_xgboost < 5500))
+                        Xpfeh_gaia_dr3.teff_xgboost < 5500,
+                        Xpfeh_gaia_dr3.mh_xgboost < -1.0,
+                        m_w1 > -0.3 - 0.006 * (5500 - Xpfeh_gaia_dr3.teff_xgboost),
+                        m_w1 > -0.01 * (5300 - Xpfeh_gaia_dr3.teff_xgboost)))
 
         # Gaia_DR3 peewee model class corresponds to
         # table catalogdb.gaia_dr3_source.
@@ -252,7 +267,7 @@ Lead contact: Alexander Ji, Rene Andrae
         Priority: three levels
         2100 if mh_xgboost <= -2.0 (TBA pending A/B test)
         2970 if -2.0 < mh_xgboost <= -1.5
-        6090 if mh_xgboost > -1.5
+        6090 if -1.5 < mh_xgboost <= -1.0
 
         Instrument: BOSS for G>13, APOGEE for G<13
         """
@@ -268,6 +283,7 @@ Lead contact: Alexander Ji, Rene Andrae
             current_mh_xgboost = output[i][1]
             current_phot_g_mean_mag = output[i][2]
 
+            # From query in build_query(), we have mh_xgboost < -1.0
             if (current_mh_xgboost <= -2.0):
                 current_priority = 2100
             elif ((current_mh_xgboost > -2.0) and
@@ -304,7 +320,7 @@ Link to paper: https://arxiv.org/abs/2302.02611 and
 Zenodo: https://doi.org/10.5281/zenodo.7599789
 Use catalogdb.xpfeh_gaia_dr3.
 
-Cut on BP<17, teff_xgboost < 5500, logg_xgboost < 4,
+Cut on BP<17, teff_xgboost < 5500, logg_xgboost < 4, mh_xgboost < -1.0
 w1mpro (from AllWise) absolute magnitude.
 
 Specifically:
@@ -314,6 +330,7 @@ with teff_xgboost on right hand side)
 BP < 17
 logg_xgboost < 4.0
 teff_xgboost < 5500
+mh_xgboost < -1.0
 M_W1 > -0.3 - 0.006 * (5500 - teff_xgboost)
 M_W1 > -0.01 * (5300 - teff_xgboost)
 
@@ -328,7 +345,7 @@ Metadata:
 Priority:
 2099 if mh_xgboost <= -2.0 (TBA pending A/B test)
 2969 if -2.0 < mh_xgboost <= -1.5
-6090 if mh_xgboost > -1.5
+6090 if -1.5 < mh_xgboost <= -1.0
 Cadence: dark_1x2
 Instrument: BOSS for G>13, APOGEE for G<13
 can_offset = True
@@ -346,6 +363,12 @@ Lead contact: Alexander Ji, Rene Andrae
 
     def build_query(self, version_id, query_region=None):
 
+        # We use the lower case variable name m_w1 so that the
+        # corresponding column name is lower case.
+        m_w1 = AllWise.w1mpro + 5 * peewee.fn.log10(Gaia_DR3.parallax / 100)
+
+        # w1mpro is of PostgreSQL numeric(5,3) type
+        # Hence, below we cast it to real so that astropy can handle it.
         query = (CatalogToGaia_DR3
                  .select(CatalogToGaia_DR3.catalogid,
                          Gaia_DR3.source_id,
@@ -353,10 +376,12 @@ Lead contact: Alexander Ji, Rene Andrae
                          Gaia_DR3.dec.alias('gaia_dr3_dec'),
                          Gaia_DR3.phot_bp_mean_mag,
                          Gaia_DR3.phot_g_mean_mag,
+                         Gaia_DR3.parallax,
                          Xpfeh_gaia_dr3.logg_xgboost,
                          Xpfeh_gaia_dr3.teff_xgboost,
                          Xpfeh_gaia_dr3.mh_xgboost,
-                         AllWise.w1mpro)
+                         AllWise.w1mpro.cast('real'),
+                         m_w1.alias('m_w1'))
                  .join(Gaia_DR3, on=(CatalogToGaia_DR3.target_id == Gaia_DR3.source_id))
                  .join(Xpfeh_gaia_dr3,
                        on=(Gaia_DR3.source_id == Xpfeh_gaia_dr3.source_id))
@@ -367,9 +392,15 @@ Lead contact: Alexander Ji, Rene Andrae
                        on=(CatalogToAllWise.target_id == AllWise.cntr))
                  .where(CatalogToGaia_DR3.version_id == version_id,
                         CatalogToGaia_DR3.best >> True,
+                        CatalogToAllWise.version_id == version_id,
+                        CatalogToAllWise.best >> True,
                         Gaia_DR3.phot_bp_mean_mag < 17,
+                        Gaia_DR3.parallax > 0,
                         Xpfeh_gaia_dr3.logg_xgboost < 4.0,
-                        Xpfeh_gaia_dr3.teff_xgboost < 5500))
+                        Xpfeh_gaia_dr3.teff_xgboost < 5500,
+                        Xpfeh_gaia_dr3.mh_xgboost < -1.0,
+                        m_w1 > -0.3 - 0.006 * (5500 - Xpfeh_gaia_dr3.teff_xgboost),
+                        m_w1 > -0.01 * (5300 - Xpfeh_gaia_dr3.teff_xgboost)))
 
         # Gaia_DR3 peewee model class corresponds to
         # table catalogdb.gaia_dr3_source.
@@ -390,7 +421,7 @@ Lead contact: Alexander Ji, Rene Andrae
         Priority: three levels
         2099 if mh_xgboost <= -2.0 (TBA pending A/B test)
         2969 if -2.0 < mh_xgboost <= -1.5
-        6090 if mh_xgboost > -1.5
+        6090 if -1.5 < mh_xgboost <= -1.0
 
         Instrument: BOSS for G>13, APOGEE for G<13
         """
@@ -406,6 +437,7 @@ Lead contact: Alexander Ji, Rene Andrae
             current_mh_xgboost = output[i][1]
             current_phot_g_mean_mag = output[i][2]
 
+            # From query in build_query(), we have mh_xgboost < -1.0
             if (current_mh_xgboost <= -2.0):
                 current_priority = 2099
             elif ((current_mh_xgboost > -2.0) and
@@ -545,6 +577,7 @@ Note: LH_MSTO5 priority 3030 Covers both vtan>200 and vtan>150.
                         Gaia_DR3.parallax > 0.5,
                         Gaia_DR3.parallax_over_error > 5,
                         Gaia_DR3.phot_rp_mean_mag < 20,
+                        Gaia_DR3.phot_g_mean_mag.is_null(False),
                         Gaia_DR3.ruwe < 1.4,
                         vtan > 150))
 
@@ -609,6 +642,8 @@ Note: LH_MSTO5 priority 3030 Covers both vtan>200 and vtan>150.
             current_phot_g_mean_mag = output[i][3]
             current_parallax = output[i][4]
 
+            # The 'where' clause in the query in build_query() ensures
+            # that current_phot_g_mean_mag and current_parallax are not null.
             current_m_g = current_phot_g_mean_mag - (10 - 5 * math.log10(current_parallax))
 
             m_g_3to5 = ((3 < current_m_g) and (current_m_g < 5))
@@ -813,6 +848,7 @@ Note: LH_MSTO5 priority 3030 Covers both vtan>200 and vtan>150.
                         Gaia_DR3.parallax > 0.5,
                         Gaia_DR3.parallax_over_error > 5,
                         Gaia_DR3.phot_rp_mean_mag < 20,
+                        Gaia_DR3.phot_g_mean_mag.is_null(False),
                         Gaia_DR3.ruwe < 1.4,
                         vtan > 150))
 
@@ -885,6 +921,8 @@ Note: LH_MSTO5 priority 3030 Covers both vtan>200 and vtan>150.
             current_phot_g_mean_mag = output[i][3]
             current_parallax = output[i][4]
 
+            # The 'where' clause in the query in build_query() ensures
+            # that current_phot_g_mean_mag and current_parallax are not null.
             current_m_g = current_phot_g_mean_mag - (10 - 5 * math.log10(current_parallax))
 
             m_g_3to5 = ((3 < current_m_g) and (current_m_g < 5))